+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 0.1.0
+ - Deploy OpenDKIM
+ - Deploy OpenDMARC
+
 # Version 0.0.1
  - Initial release